Move constants to their related modules.
