+++ bittensor/axon.py
@@ -45,7 +45,6 @@ from substrateinterface import Keypair
 
 import bittensor
 from bittensor.utils.axon_utils import allowed_nonce_window_ns, calculate_diff_seconds
-from bittensor.constants import V_7_2_0
 from bittensor.errors import (
     BlacklistedException,
     InvalidRequestNameError,
@@ -59,6 +58,8 @@ from bittensor.errors import (
 )
 from bittensor.threadpool import PriorityThreadPoolExecutor
 from bittensor.utils import networking
+
+V_7_2_0 = 7002000
 
 
 class FastAPIThreadedServer(uvicorn.Server):
