+++ bittensor/dendrite.py
@@ -1,36 +1,43 @@
 # The MIT License (MIT)
-# Copyright © 2021 Yuma Rao
-# Copyright © 2022 Opentensor Foundation
-# Copyright © 2023 Opentensor Technologies Inc
-
+# Copyright © 2024 Opentensor Foundation
+#
 # Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
 # documentation files (the “Software”), to deal in the Software without restriction, including without limitation
 # the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
 # and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-
+#
 # The above copyright notice and this permission notice shall be included in all copies or substantial portions of
 # the Software.
-
+#
 # THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
 # THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
 # THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 
-# Standard Library
 from __future__ import annotations
+
 import asyncio
 import time
-from typing import Optional, List, Union, AsyncGenerator, Any
 import uuid
+from typing import Any, AsyncGenerator, Dict, List, Optional, Union, Type
 
-# 3rd Party
 import aiohttp
 
-# Application
 import bittensor
-from bittensor.constants import DENDRITE_ERROR_MAPPING, DENDRITE_DEFAULT_ERROR
 from bittensor.utils.registration import torch, use_torch
+
+DENDRITE_ERROR_MAPPING: Dict[Type[Exception], tuple] = {
+    aiohttp.ClientConnectorError: ("503", "Service unavailable"),
+    asyncio.TimeoutError: ("408", "Request timeout"),
+    aiohttp.ClientResponseError: (None, "Client response error"),
+    aiohttp.ClientPayloadError: ("400", "Payload error"),
+    aiohttp.ClientError: ("500", "Client error"),
+    aiohttp.ServerTimeoutError: ("504", "Server timeout error"),
+    aiohttp.ServerDisconnectedError: ("503", "Service disconnected"),
+    aiohttp.ServerConnectionError: ("503", "Service connection error"),
+}
+DENDRITE_DEFAULT_ERROR = ("422", "Failed to parse response")
 
 
 class DendriteMixin:
